Addresses GH #21918
removed test_operations.py

fixed syntax for python 2.7

fix comments from commit
