+++ pandas/tests/indexes/multi/test_format.py
@@ -100,11 +100,6 @@ def test_repr_roundtrip():
     tm.assert_index_equal(result, mi_u, exact=True)
 
 
-def test_str():
-    # tested elsewhere
-    pass
-
-
 def test_unicode_string_with_unicode():
     d = {"a": [u("\u05d0"), 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]}
     idx = pd.DataFrame(d).set_index(["a", "b"]).index
