+++ pandas/tests/indexes/multi/test_sorting.py
@@ -254,3 +254,9 @@ def test_remove_unused_nan(level0, level1):
     tm.assert_index_equal(result, mi)
     for level in 0, 1:
         assert('unused' not in result.levels[level])
+
+
+def test_argsort(idx):
+    result = idx.argsort()
+    expected = idx.values.argsort()
+    tm.assert_numpy_array_equal(result, expected)
